Move add-on modules into addons folder
